+++ packages/flate/lib/src/widget.dart
@@ -1,5 +1,5 @@
 export 'widget/component.dart';
 export 'widget/container.dart';
-export 'widget/should_not_rebuild.dart';
+export 'widget/reduce_rebuild.dart';
 export 'widget/activity.dart';
 export 'widget/action.dart';